Refactor file repositories into separate package in attempt at clarity
