+++ src/main/java/gov/cdc/izgateway/xform/repository/file/AccessControlFileRepository.java
@@ -1,23 +1,22 @@
-package gov.cdc.izgateway.xform.repository;
+package gov.cdc.izgateway.xform.repository.file;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import gov.cdc.izgateway.xform.model.AccessControl;
-import gov.cdc.izgateway.xform.model.GroupRoleMapping;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Repository;
 
 import java.util.LinkedHashSet;
 
 @Repository
-public class GroupRoleMappingFileRepository extends GenericFileRepository<GroupRoleMapping> {
+public class AccessControlFileRepository extends GenericFileRepository<AccessControl> {
 
-    @Value("${xform.configurations.group-role-mapping}")
+    @Value("${xform.configurations.access-control}")
     protected void setFilePath(String filePath) {
         this.filePath = filePath;
     }
 
     @Override
-    protected TypeReference<LinkedHashSet<GroupRoleMapping>> getTypeReference() {
+    protected TypeReference<LinkedHashSet<AccessControl>> getTypeReference() {
         return new TypeReference<>() {};
     }
 }
