+++ src/main/java/gov/cdc/izgateway/xform/repository/file/MappingFileRepository.java
@@ -1,22 +1,23 @@
-package gov.cdc.izgateway.xform.repository;
+package gov.cdc.izgateway.xform.repository.file;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import gov.cdc.izgateway.xform.model.AccessControl;
+import gov.cdc.izgateway.xform.model.Mapping;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Repository;
 
 import java.util.LinkedHashSet;
 
 @Repository
-public class AccessControlFileRepository extends GenericFileRepository<AccessControl> {
+public class MappingFileRepository extends GenericFileRepository<Mapping> {
 
-    @Value("${xform.configurations.access-control}")
+    @Value("${xform.configurations.mappings}")
     protected void setFilePath(String filePath) {
         this.filePath = filePath;
     }
 
+
     @Override
-    protected TypeReference<LinkedHashSet<AccessControl>> getTypeReference() {
+    protected TypeReference<LinkedHashSet<Mapping>> getTypeReference() {
         return new TypeReference<>() {};
     }
 }
