+++ src/main/java/gov/cdc/izgateway/xform/repository/file/OrganizationFileRepository.java
@@ -1,7 +1,7 @@
-package gov.cdc.izgateway.xform.repository;
+package gov.cdc.izgateway.xform.repository.file;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import gov.cdc.izgateway.xform.model.Pipeline;
+import gov.cdc.izgateway.xform.model.Organization;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Repository;
@@ -10,16 +10,16 @@ import java.util.LinkedHashSet;
 
 @Repository
 @ConditionalOnProperty(name = "xform.repository.type", havingValue = "file", matchIfMissing = true)
-public class PipelineFileRepository extends GenericFileRepository<Pipeline> {
+public class OrganizationFileRepository extends GenericFileRepository<Organization> {
 
-    @Value("${xform.configurations.pipelines}")
+    @Value("${xform.configurations.organizations}")
     protected void setFilePath(String filePath) {
         this.filePath = filePath;
     }
 
+
     @Override
-    protected TypeReference<LinkedHashSet<Pipeline>> getTypeReference() {
+    protected TypeReference<LinkedHashSet<Organization>> getTypeReference() {
         return new TypeReference<>() {};
     }
-
 }
