+++ src/main/java/gov/cdc/izgateway/xform/repository/file/PipelineFileRepository.java
@@ -1,26 +1,25 @@
-package gov.cdc.izgateway.xform.repository;
+package gov.cdc.izgateway.xform.repository.file;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import gov.cdc.izgateway.xform.model.Organization;
+import gov.cdc.izgateway.xform.model.Pipeline;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Repository;
 
 import java.util.LinkedHashSet;
 
 @Repository
 @ConditionalOnProperty(name = "xform.repository.type", havingValue = "file", matchIfMissing = true)
-public class OrganizationFileRepository extends GenericFileRepository<Organization> {
+public class PipelineFileRepository extends GenericFileRepository<Pipeline> {
 
-    @Value("${xform.configurations.organizations}")
+    @Value("${xform.configurations.pipelines}")
     protected void setFilePath(String filePath) {
         this.filePath = filePath;
     }
 
-
     @Override
-    protected TypeReference<LinkedHashSet<Organization>> getTypeReference() {
+    protected TypeReference<LinkedHashSet<Pipeline>> getTypeReference() {
         return new TypeReference<>() {};
     }
+
 }
