+++ src/main/java/gov/cdc/izgateway/xform/repository/file/SolutionFileRepository.java
@@ -1,23 +1,22 @@
-package gov.cdc.izgateway.xform.repository;
+package gov.cdc.izgateway.xform.repository.file;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import gov.cdc.izgateway.xform.model.Mapping;
+import gov.cdc.izgateway.xform.model.Solution;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Repository;
 
 import java.util.LinkedHashSet;
 
 @Repository
-public class MappingFileRepository extends GenericFileRepository<Mapping> {
+public class SolutionFileRepository extends GenericFileRepository<Solution> {
 
-    @Value("${xform.configurations.mappings}")
+    @Value("${xform.configurations.solutions}")
     protected void setFilePath(String filePath) {
         this.filePath = filePath;
     }
 
-
     @Override
-    protected TypeReference<LinkedHashSet<Mapping>> getTypeReference() {
+    protected TypeReference<LinkedHashSet<Solution>> getTypeReference() {
         return new TypeReference<>() {};
     }
 }
